treewide: Align networking code with net_mgmt API change
The net_mgmt event representation changed from 32-bit to 64-bit value,
therefore update all net_mgmt callbacks to align with this change.

+++ samples/cellular/nrf_cloud_multi_service/src/provisioning_support.c
@@ -171,7 +171,7 @@ static K_WORK_DEFINE(start_provisioning_work, start_provisioning_work_fn);
 static struct net_mgmt_event_callback l4_callback;
 
 static void l4_event_handler(struct net_mgmt_event_callback *cb,
-			     uint32_t event, struct net_if *iface)
+			     uint64_t event, struct net_if *iface)
 {
 	if (event == NET_EVENT_L4_CONNECTED) {
 		/* Mark network as up. */
